plantguard-backend/plantguard_ai.py

import sys
import json
import os
import numpy as np
import tensorflow as tf
from PIL import Image

def load_model():
    # Placeholder for actual model loading
    # In a real scenario, you'd load a pre-trained model
    model = tf.keras.models.load_model('path/to/your/trained/model.h5')
    return model

def preprocess_image(image_path):
    img = Image.open(image_path)
    img = img.resize((224, 224))  # Adjust size as per your model
    img_array = np.array(img) / 255.0
    img_array = np.expand_dims(img_array, axis=0)
    return img_array

def analyze_plant_disease(image_path):
    model = load_model()
    processed_image = preprocess_image(image_path)
    
    # Predict
    predictions = model.predict(processed_image)
    
    # Placeholder logic - replace with actual model interpretation
    disease_classes = [
        'Healthy', 
        'Powdery Mildew', 
        'Leaf Spot', 
        'Root Rot'
    ]
    
    predicted_class = disease_classes[np.argmax(predictions)]
    confidence = float(np.max(predictions))

    # Generate recommendations based on disease
    recommendations = {
        'Healthy': ['Continue current care routine'],
        'Powdery Mildew': [
            'Isolate the plant', 
            'Reduce humidity', 
            'Apply fungicide'
        ],
        'Leaf Spot': [
            'Remove affected leaves', 
            'Improve air circulation', 
            'Avoid overhead watering'
        ],
        'Root Rot': [
            'Reduce watering', 
            'Improve drainage', 
            'Repot with fresh soil'
        ]
    }

    return {
        'species': 'Unknown Plant',
        'disease': predicted_class,
        'confidence': confidence,
        'recommendations': recommendations.get(predicted_class, [])
    }

def main():
    if len(sys.argv) < 2:
        print(json.dumps({
            'error': 'No image path provided'
        }))
        sys.exit(1)

    image_path = sys.argv[1]
    
    try:
        result = analyze_plant_disease(image_path)
        print(json.dumps(result))
    except Exception as e:
        print(json.dumps({
            'error': str(e)
        }))
        sys.exit(1)

if __name__ == '__main__':
    main()
